requirements.txt
streamlit
google-generativeai
python-dotenv
pyttsx3
speechrecognition
pdfplumber
duckduckgo-search